backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
pydantic>=2.0.0
